+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+6.3.0
+-----
+
+- Fix IMAP IDLE response processing for some mail servers (#67)
+- Exit with a critical error when options are missing (#68)
+- Add IMAP responses to debug logging
+- Add `smtp` option `skip_certificate_verification`
+- Add `kafka` option `skip_certificate_verification`
+- Suppress `mailparser` logging output
+- Suppress `msgconvert` warnings
+
 6.2.2
 -----